refactor: 메뉴 관련 테스트 코드 역할 분리
