+++ src/test/java/profect/eatcloud/Domain/manager/controller/ManagerMenuControllerTest.java
@@ -1,0 +1,82 @@
+package profect.eatcloud.Domain.manager.controller;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.http.MediaType;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.setup.MockMvcBuilders;
+import profect.eatcloud.Domain.Manager.Controller.ManagerController;
+import profect.eatcloud.Domain.Manager.Service.ManagerService;
+import profect.eatcloud.Domain.Store.Controller.MenuController;
+import profect.eatcloud.Domain.Store.Dto.MenuRequestDto;
+import profect.eatcloud.Domain.Store.Entity.Menu;
+import profect.eatcloud.Domain.Store.Service.MenuService;
+
+import java.math.BigDecimal;
+import java.util.UUID;
+
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.eq;
+import static org.mockito.Mockito.when;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+
+@ExtendWith(MockitoExtension.class)
+class ManagerMenuControllerTest {
+
+    @InjectMocks
+    private ManagerController managerController;
+
+    @Mock
+    private ManagerService managerService;
+
+    private MockMvc mockMvc;
+
+    private final ObjectMapper objectMapper = new ObjectMapper();
+
+    @BeforeEach
+    void setup() {
+        mockMvc = MockMvcBuilders.standaloneSetup(managerController).build();
+    }
+
+    @Test
+    void 메뉴등록_성공() throws Exception {
+        // given
+        UUID storeId = UUID.randomUUID();
+        MenuRequestDto dto = new MenuRequestDto();
+        dto.setMenuNum(1);
+        dto.setMenuName("불고기 덮밥");
+        dto.setMenuCategoryCode("KOREAN");
+        dto.setPrice(new BigDecimal("8500"));
+        dto.setDescription("불고기와 밥 구성");
+        dto.setIsAvailable(true);
+        dto.setImageUrl("image.jpg");
+
+        Menu savedMenu = Menu.builder()
+                .id(UUID.randomUUID())
+                .menuName(dto.getMenuName())
+                .price(dto.getPrice())
+                .build();
+
+        when(managerService.createMenu(eq(storeId), any(MenuRequestDto.class)))
+                .thenReturn(savedMenu);
+
+        // when & then
+        mockMvc.perform(post("/api/v1/manager/stores/{storeId}/menus", storeId)
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .content(objectMapper.writeValueAsString(dto)))
+                .andExpect(status().isOk())
+                .andExpect(jsonPath("$.menuName").value("불고기 덮밥"));
+    }
+
+
+}
+
+
