--- src/test/java/profect/eatcloud/Domain/store/service/MenuServiceTest.java
@@ -1,211 +1,0 @@
-package profect.eatcloud.Domain.store.service;
-
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import profect.eatcloud.Domain.Store.Dto.MenuRequestDto;
-import profect.eatcloud.Domain.Store.Entity.Menu;
-import profect.eatcloud.Domain.Store.Entity.Store;
-import profect.eatcloud.Domain.Store.Repository.MenuRepository_min;
-import profect.eatcloud.Domain.Store.Repository.StoreRepository_min;
-import profect.eatcloud.Domain.Store.Service.MenuService;
-
-import java.math.BigDecimal;
-import java.util.Optional;
-import java.util.UUID;
-
-import static org.junit.jupiter.api.Assertions.*;
-import static org.mockito.Mockito.when;
-
-import org.junit.jupiter.api.BeforeEach;
-import org.mockito.ArgumentCaptor;
-
-import static org.mockito.Mockito.*;
-
-class MenuServiceTest {
-
-    private MenuService menuService;
-    private MenuRepository_min menuRepository;
-    private StoreRepository_min storeRepository;
-
-    @BeforeEach
-    void setUp() {
-        menuRepository = mock(MenuRepository_min.class);
-        storeRepository = mock(StoreRepository_min.class);
-        menuService = new MenuService(menuRepository, storeRepository);
-    }
-
-    @Test
-    void 메뉴가_정상적으로_등록된다() {
-        // given
-        UUID storeId = UUID.randomUUID();
-        Store store = Store.builder().storeId(storeId).build();
-        when(storeRepository.findById(storeId)).thenReturn(Optional.of(store));
-
-        MenuRequestDto dto = new MenuRequestDto();
-        dto.setMenuNum(1);
-        dto.setMenuName("테스트메뉴");
-        dto.setMenuCategoryCode("KOREAN");
-        dto.setPrice(new BigDecimal("12000"));
-        dto.setDescription("맛있는 한식");
-        dto.setIsAvailable(true);
-        dto.setImageUrl("image-url");
-
-        // when
-        menuService.createMenu(storeId, dto);
-
-        // then
-        ArgumentCaptor<Menu> captor = ArgumentCaptor.forClass(Menu.class);
-        verify(menuRepository, times(1)).save(captor.capture());
-
-        Menu savedMenu = captor.getValue();
-        assertEquals("테스트메뉴", savedMenu.getMenuName());
-        assertEquals("KOREAN", savedMenu.getMenuCategoryCode());
-    }
-
-    @Test
-    void 존재하지_않는_가게일_경우_예외가_발생한다() {
-        // given
-        UUID storeId = UUID.randomUUID();
-        when(storeRepository.findById(storeId)).thenReturn(Optional.empty());
-
-        MenuRequestDto dto = new MenuRequestDto();
-
-        // when + then
-        IllegalArgumentException exception = assertThrows(IllegalArgumentException.class, () -> {
-            menuService.createMenu(storeId, dto);
-        });
-
-        assertEquals("존재하지 않는 매장입니다", exception.getMessage());
-    }
-
-    @Test
-    void 가격이_음수일_경우_예외가_발생한다() {
-        // given
-        UUID storeId = UUID.randomUUID();
-        Store store = Store.builder().storeId(storeId).build();
-        when(storeRepository.findById(storeId)).thenReturn(Optional.of(store));
-
-        MenuRequestDto dto = new MenuRequestDto();
-        dto.setMenuName("제육덮밥");
-        dto.setPrice(new BigDecimal("-1000"));
-
-        // when + then
-        IllegalArgumentException exception = assertThrows(IllegalArgumentException.class, () -> {
-            menuService.createMenu(storeId, dto);
-        });
-
-        assertEquals("가격은 0 이상이어야 합니다.", exception.getMessage());
-    }
-
-    @Test
-    @DisplayName("메뉴 이름이 비어 있으면 예외가 발생한다")
-    void 메뉴이름이_없으면_예외발생() {
-        // given
-        UUID storeId = UUID.randomUUID();
-        Store store = Store.builder().storeId(storeId).build();
-        when(storeRepository.findById(storeId)).thenReturn(Optional.of(store));
-
-        MenuRequestDto dto = new MenuRequestDto();
-        dto.setMenuName(""); // 비어 있는 이름
-        dto.setMenuCategoryCode("KOREAN");
-        dto.setPrice(new BigDecimal("12000"));
-        dto.setIsAvailable(true);
-
-        // when & then
-        IllegalArgumentException exception = assertThrows(IllegalArgumentException.class, () -> {
-            menuService.createMenu(storeId, dto);
-        });
-
-        assertEquals("메뉴 이름은 필수입니다", exception.getMessage());
-    }
-
-    @Test
-    @DisplayName("정상적인 메뉴 저장 시 모든 필드가 올바르게 저장된다")
-    void 정상메뉴저장_모든필드검증() {
-        // given
-        UUID storeId = UUID.randomUUID();
-        Store store = Store.builder().storeId(storeId).build();
-        when(storeRepository.findById(storeId)).thenReturn(Optional.of(store));
-
-        MenuRequestDto dto = new MenuRequestDto();
-        dto.setMenuNum(10);
-        dto.setMenuName("삼겹살 정식");
-        dto.setMenuCategoryCode("KOREAN");
-        dto.setPrice(new BigDecimal("13000"));
-        dto.setDescription("고소한 삼겹살과 밥, 반찬 구성");
-        dto.setIsAvailable(null); // 명시적 null
-        dto.setImageUrl("image-url.png");
-
-        // when
-        menuService.createMenu(storeId, dto);
-
-        // then
-        ArgumentCaptor<Menu> captor = ArgumentCaptor.forClass(Menu.class);
-        verify(menuRepository).save(captor.capture());
-
-        Menu saved = captor.getValue();
-        assertEquals(10, saved.getMenuNum());
-        assertEquals("삼겹살 정식", saved.getMenuName());
-        assertEquals("KOREAN", saved.getMenuCategoryCode());
-        assertEquals(new BigDecimal("13000"), saved.getPrice());
-        assertEquals("고소한 삼겹살과 밥, 반찬 구성", saved.getDescription());
-        assertEquals("image-url.png", saved.getImageUrl());
-        assertTrue(saved.getIsAvailable()); // 기본값 true 적용 확인
-    }
-
-    @Test
-    void 같은가게에서_중복된_메뉴번호로_등록하면_예외가_발생한다() {
-        // given
-        UUID storeId = UUID.randomUUID();
-        Store store = Store.builder().storeId(storeId).build();
-
-        when(storeRepository.findById(storeId)).thenReturn(Optional.of(store));
-        when(menuRepository.existsByStoreAndMenuNum(store, 1)).thenReturn(true); // 중복된 번호
-
-        MenuRequestDto dto = new MenuRequestDto();
-        dto.setMenuNum(1);
-        dto.setMenuName("중복메뉴");
-        dto.setPrice(new BigDecimal("10000"));
-
-        // when + then
-        IllegalArgumentException e = assertThrows(IllegalArgumentException.class, () -> {
-            menuService.createMenu(storeId, dto);
-        });
-
-        assertEquals("해당 메뉴 번호는 이미 존재합니다", e.getMessage());
-    }
-
-    @Test
-    void 존재하지_않는_메뉴를_삭제하면_예외발생() {
-        // given
-        UUID menuId = UUID.randomUUID();
-        when(menuRepository.findById(menuId)).thenReturn(Optional.empty());
-
-        // when & then
-        IllegalArgumentException e = assertThrows(IllegalArgumentException.class, () -> {
-            menuService.deleteMenu(menuId);
-        });
-
-        assertEquals("존재하지 않는 메뉴입니다", e.getMessage());
-    }
-
-    @Test
-    void 메뉴_삭제가_정상적으로_동작한다() {
-        // given
-        UUID menuId = UUID.randomUUID();
-        Menu menu = Menu.builder().id(menuId).build();
-        when(menuRepository.findById(menuId)).thenReturn(Optional.of(menu));
-
-        // when
-        menuService.deleteMenu(menuId);
-
-        // then
-        verify(menuRepository, times(1)).deleteById(menuId);
-    }
-
-
-
-
-
-
-}
